# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: SentraTest API
  version: 1.0.0
  description: REST API for SentraTest – AI-powered test case generation platform.

servers:
  - url: http://localhost:4000
    description: Local development

tags:
  - name: Users
    description: User management
  - name: Chats
    description: Chat sessions and messages
  - name: Test Cases
    description: AI-generated test cases
  - name: Files
    description: Uploaded file metadata
  - name: Exports
    description: Export history

paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  # ── Users ──────────────────────────────────────────────────────────────────
  /api/users:
    post:
      tags: [Users]
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'

  /api/users/{id}:
    get:
      tags: [Users]
      summary: Get user by ID
      parameters:
        - $ref: '#/components/parameters/UUIDParam'
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Users]
      summary: Update user fields
      parameters:
        - $ref: '#/components/parameters/UUIDParam'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Updated
        '400':
          $ref: '#/components/responses/BadRequest'
    delete:
      tags: [Users]
      summary: Delete user
      parameters:
        - $ref: '#/components/parameters/UUIDParam'
      responses:
        '204':
          description: Deleted

  # ── Chats ──────────────────────────────────────────────────────────────────
  /api/chats:
    post:
      tags: [Chats]
      summary: Create a new chat (optionally with first message)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChatRequest'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdResponse'
          description: Chat created
        '400':
          $ref: '#/components/responses/BadRequest'
    get:
      tags: [Chats]
      summary: List chats for a user
      parameters:
        - $ref: '#/components/parameters/UserIdQuery'
        - $ref: '#/components/parameters/LimitQuery'
        - $ref: '#/components/parameters/OffsetQuery'
      responses:
        '200':
          description: Paginated chat list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedChats'

  /api/chats/{id}:
    delete:
      tags: [Chats]
      summary: Delete chat and its messages
      parameters:
        - $ref: '#/components/parameters/UUIDParam'
      responses:
        '204':
          description: Deleted

  /api/chats/{chatId}/messages:
    post:
      tags: [Chats]
      summary: Add a message to a chat
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMessageRequest'
      responses:
        '201':
          description: Message added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
    get:
      tags: [Chats]
      summary: List messages in a chat
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/LimitQuery'
        - $ref: '#/components/parameters/OffsetQuery'
      responses:
        '200':
          description: Paginated message list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedMessages'

  # ── Test Cases ─────────────────────────────────────────────────────────────
  /api/test-cases:
    post:
      tags: [Test Cases]
      summary: Store a generated test case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTestCaseRequest'
      responses:
        '201':
          description: Test case stored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
    get:
      tags: [Test Cases]
      summary: List test cases for a user (paginated)
      parameters:
        - $ref: '#/components/parameters/UserIdQuery'
        - $ref: '#/components/parameters/LimitQuery'
        - $ref: '#/components/parameters/OffsetQuery'
      responses:
        '200':
          description: Paginated list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTestCases'

  /api/test-cases/search:
    get:
      tags: [Test Cases]
      summary: Full-text search test cases
      parameters:
        - $ref: '#/components/parameters/UserIdQuery'
        - name: q
          in: query
          required: true
          description: Search keyword
          schema:
            type: string
        - $ref: '#/components/parameters/LimitQuery'
        - $ref: '#/components/parameters/OffsetQuery'
      responses:
        '200':
          description: Matching test cases with relevance score
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTestCases'
        '400':
          $ref: '#/components/responses/BadRequest'

  # ── Files ──────────────────────────────────────────────────────────────────
  /api/files:
    post:
      tags: [Files]
      summary: Record a file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFileRequest'
      responses:
        '201':
          description: File recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdResponse'
    get:
      tags: [Files]
      summary: List uploaded files for a user
      parameters:
        - $ref: '#/components/parameters/UserIdQuery'
        - $ref: '#/components/parameters/LimitQuery'
        - $ref: '#/components/parameters/OffsetQuery'
      responses:
        '200':
          description: Paginated file list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedFiles'

  /api/files/{id}:
    delete:
      tags: [Files]
      summary: Delete uploaded file record
      parameters:
        - $ref: '#/components/parameters/UUIDParam'
      responses:
        '204':
          description: Deleted

  # ── Exports ────────────────────────────────────────────────────────────────
  /api/exports:
    post:
      tags: [Exports]
      summary: Log an export event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExportRequest'
      responses:
        '201':
          description: Export logged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
    get:
      tags: [Exports]
      summary: List export history for a user
      parameters:
        - $ref: '#/components/parameters/UserIdQuery'
        - $ref: '#/components/parameters/LimitQuery'
        - $ref: '#/components/parameters/OffsetQuery'
      responses:
        '200':
          description: Paginated export history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedExports'

components:
  parameters:
    UUIDParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    UserIdQuery:
      name: userId
      in: query
      required: true
      schema:
        type: string
        format: uuid
    LimitQuery:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
    OffsetQuery:
      name: offset
      in: query
      schema:
        type: integer
        default: 0

  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict (e.g. duplicate email)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    IdResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid

    ErrorResponse:
      type: object
      properties:
        error:
          type: string

    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        username:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    CreateUserRequest:
      type: object
      required: [email, username, passwordHash]
      properties:
        email:
          type: string
          format: email
        username:
          type: string
        passwordHash:
          type: string

    UpdateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        username:
          type: string

    CreateChatRequest:
      type: object
      required: [userId]
      properties:
        userId:
          type: string
          format: uuid
        title:
          type: string
        initialMessage:
          type: string

    AddMessageRequest:
      type: object
      required: [userId, role, content]
      properties:
        userId:
          type: string
          format: uuid
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string

    CreateTestCaseRequest:
      type: object
      required: [userId, input, output]
      properties:
        userId:
          type: string
          format: uuid
        input:
          type: string
        output:
          type: object
        method:
          type: string
        type:
          type: string
          enum: [blackbox, whitebox]

    CreateFileRequest:
      type: object
      required: [userId, filename, path]
      properties:
        userId:
          type: string
          format: uuid
        filename:
          type: string
        path:
          type: string
        type:
          type: string
        size:
          type: integer

    CreateExportRequest:
      type: object
      required: [userId, format, testCaseCount, filePath]
      properties:
        userId:
          type: string
          format: uuid
        format:
          type: string
          enum: [CSV, JSON, Excel, PDF]
        testCaseCount:
          type: integer
        filePath:
          type: string

    PaginatedChats:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
        limit:
          type: integer
        offset:
          type: integer

    PaginatedMessages:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
        limit:
          type: integer
        offset:
          type: integer

    PaginatedTestCases:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
        limit:
          type: integer
        offset:
          type: integer

    PaginatedFiles:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
        limit:
          type: integer
        offset:
          type: integer

    PaginatedExports:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
        limit:
          type: integer
        offset:
          type: integer